lexicon annotations support added reflecting last changes in Lexicon TCF schema.

+++ src/main/java/eu/clarin/weblicht/wlfxb/lx/api/Frequency.java
@@ -20,7 +20,11 @@
  */
 package eu.clarin.weblicht.wlfxb.lx.api;
 
+/**
+ * @author Yana Panchenko
+ *
+ */
 public interface Frequency {
 
-    public int getValue();
+    public double getValue();
 }
